implement GroupByValidatorTest
